+++ lib/claide/help.rb
@@ -18,9 +18,18 @@ module CLAide
     #
     attr_reader :error_message
 
-    # @return [Bool] Whether the error message should be colorized.
+    # @return [Bool] Whether the error message should use ANSI codes to
+    #         prettify output.
     #
-    attr_reader  :colorize
+    attr_reader  :ansi_output
+    alias_method :ansi_output?, :ansi_output
+
+    def colorize
+      warn "[!] The use of `CLAide::Help#colorize` has been " \
+           "deprecated. Use `CLAide::Help#ansi_output` instead. " \
+           "(Called from: #{caller.first})"
+      ansi_output
+    end
     alias_method :colorize?, :colorize
 
     # @param [String] banner @see banner
@@ -30,10 +39,10 @@ module CLAide
     #        terminate the program with, will be set to `1`, otherwise a {Help}
     #        exception is treated as not being a real error and exits with `0`.
     #
-    def initialize(banner, error_message = nil, colorize = false)
+    def initialize(banner, error_message = nil, ansi_output = false)
       @banner = banner
       @error_message = error_message
-      @colorize = colorize
+      @ansi_output = ansi_output
       @exit_status = @error_message.nil? ? 0 : 1
     end
 
@@ -43,8 +52,14 @@ module CLAide
     def formatted_error_message
       if error_message
         message = "[!] #{error_message}"
-        colorize? ? message.red : message
+        prettify_error_message(message)
       end
+    end
+
+    # @return [String]
+    #
+    def prettify_error_message(message)
+      ansi_output? ? message.red : message
     end
 
     # @return [String] The optional error message, combined with the help